feat: limitação de cadastro para turmas com 45 vagas, server responde que não há vagas

+++ server/src/routes/eletivas.ts
@@ -71,40 +71,43 @@ export async function EletivasRoutes(app: FastifyInstance) {
         if (pivo[0].classesId === classId) {
           return res.status(500).send("Usuário já cadastrado nessa disciplina");
         } else if (classFromPivo.diaDaSemana === turma.diaDaSemana) {
-          await prisma.alunosMatriculados.delete({
-            where: {
-              id: pivo[0].id,
-            },
-          });
+          if (turma.quantidadeDeAlunos <= 44) {
+            await prisma.alunosMatriculados.delete({
+              where: {
+                id: pivo[0].id,
+              },
+            });
 
-          await prisma.classes.update({
-            where: {
-              id: classFromPivo.id,
-            },
-            data: {
-              quantidadeDeAlunos: classFromPivo.quantidadeDeAlunos - 1,
-            },
-          });
+            await prisma.classes.update({
+              where: {
+                id: classFromPivo.id,
+              },
+              data: {
+                quantidadeDeAlunos: classFromPivo.quantidadeDeAlunos - 1,
+              },
+            });
 
-          const novaMatricula = await prisma.alunosMatriculados.create({
-            data: {
-              classesId: turma.id,
-              studentId: user.matricula,
-            },
-          });
+            const novaMatricula = await prisma.alunosMatriculados.create({
+              data: {
+                classesId: turma.id,
+                studentId: user.matricula,
+              },
+            });
 
-          await prisma.classes.update({
-            where: {
-              id: turma.id,
-            },
-            data: {
-              quantidadeDeAlunos: turma.quantidadeDeAlunos + 1,
-            },
-          });
+            await prisma.classes.update({
+              where: {
+                id: turma.id,
+              },
+              data: {
+                quantidadeDeAlunos: turma.quantidadeDeAlunos + 1,
+              },
+            });
 
-          return res
-            .status(200)
-            .send(`Usuário cadastrado na turma ${novaMatricula}`);
+            return res
+              .status(200)
+              .send(`Usuário cadastrado na turma ${novaMatricula}`);
+          }
+          res.status(500).send("Não há vagas disponíveis");
         }
 
         const novaMatricula = await prisma.alunosMatriculados.create({
@@ -139,73 +142,79 @@ export async function EletivasRoutes(app: FastifyInstance) {
           },
         });
         if (firstClassFromPivo.diaDaSemana === turma.diaDaSemana) {
-          await prisma.alunosMatriculados.delete({
-            where: {
-              id: pivo[0].id,
-            },
-          });
+          if (turma.quantidadeDeAlunos <= 44) {
+            await prisma.alunosMatriculados.delete({
+              where: {
+                id: pivo[0].id,
+              },
+            });
 
-          await prisma.classes.update({
-            where: {
-              id: firstClassFromPivo.id,
-            },
-            data: {
-              quantidadeDeAlunos: firstClassFromPivo.quantidadeDeAlunos - 1,
-            },
-          });
+            await prisma.classes.update({
+              where: {
+                id: firstClassFromPivo.id,
+              },
+              data: {
+                quantidadeDeAlunos: firstClassFromPivo.quantidadeDeAlunos - 1,
+              },
+            });
 
-          const novaMatricula = await prisma.alunosMatriculados.create({
-            data: {
-              classesId: turma.id,
-              studentId: user.matricula,
-            },
-          });
+            const novaMatricula = await prisma.alunosMatriculados.create({
+              data: {
+                classesId: turma.id,
+                studentId: user.matricula,
+              },
+            });
 
-          await prisma.classes.update({
-            where: {
-              id: turma.id,
-            },
-            data: {
-              quantidadeDeAlunos: turma.quantidadeDeAlunos + 1,
-            },
-          });
-          return res
-            .status(200)
-            .send(`Aluno cadastrado na nova turma ${novaMatricula}`);
-        }else if(secondClassFromPivo.diaDaSemana === turma.diaDaSemana){
-          await prisma.alunosMatriculados.delete({
-            where: {
-              id: pivo[0].id,
-            },
-          });
+            await prisma.classes.update({
+              where: {
+                id: turma.id,
+              },
+              data: {
+                quantidadeDeAlunos: turma.quantidadeDeAlunos + 1,
+              },
+            });
+            return res
+              .status(200)
+              .send(`Aluno cadastrado na nova turma ${novaMatricula}`);
+          }
+          res.status(500).send("Não há vagas disponíveis");
+        } else if (secondClassFromPivo.diaDaSemana === turma.diaDaSemana) {
+          if (turma.quantidadeDeAlunos <= 44) {
+            await prisma.alunosMatriculados.delete({
+              where: {
+                id: pivo[0].id,
+              },
+            });
 
-          await prisma.classes.update({
-            where: {
-              id: secondClassFromPivo.id,
-            },
-            data: {
-              quantidadeDeAlunos: secondClassFromPivo.quantidadeDeAlunos - 1,
-            },
-          });
+            await prisma.classes.update({
+              where: {
+                id: secondClassFromPivo.id,
+              },
+              data: {
+                quantidadeDeAlunos: secondClassFromPivo.quantidadeDeAlunos - 1,
+              },
+            });
 
-          const novaMatricula = await prisma.alunosMatriculados.create({
-            data: {
-              classesId: turma.id,
-              studentId: user.matricula,
-            },
-          });
+            const novaMatricula = await prisma.alunosMatriculados.create({
+              data: {
+                classesId: turma.id,
+                studentId: user.matricula,
+              },
+            });
 
-          await prisma.classes.update({
-            where: {
-              id: turma.id,
-            },
-            data: {
-              quantidadeDeAlunos: turma.quantidadeDeAlunos + 1,
-            },
-          });
-          return res
-            .status(200)
-            .send(`Aluno cadastrado na nova turma ${novaMatricula}`);
+            await prisma.classes.update({
+              where: {
+                id: turma.id,
+              },
+              data: {
+                quantidadeDeAlunos: turma.quantidadeDeAlunos + 1,
+              },
+            });
+            return res
+              .status(200)
+              .send(`Aluno cadastrado na nova turma ${novaMatricula}`);
+          }
+          res.status(500).send("Não há vagas disponíveis");
         }
 
         await prisma.alunosMatriculados.delete({
